Added confirm message when either delete buttons is pressed. Also change the state of delete selected button to disable after the delete action is complete.

+++ myform.py
@@ -123,16 +123,20 @@ class Feedback:
         
 
     def delete_all_list(self):
-        self.db.sql_do('DELETE FROM {}'.format(self.__t))
-        self.db.sql_do('VACUUM')
-        self.listbox_result.delete(0, 'end')
-        self.text_show_comments.config(state = 'normal')
-        self.text_show_comments.delete(1.0, 'end')
-        self.text_show_comments.config(state = 'disabled')
+        if messagebox.askokcancel("Verify", "Ok to delete all records?"):
+            self.db.sql_do('DELETE FROM {}'.format(self.__t))
+            self.db.sql_do('VACUUM')
+            self.listbox_result.delete(0, 'end')
+            self.text_show_comments.config(state = 'normal')
+            self.text_show_comments.delete(1.0, 'end')
+            self.text_show_comments.config(state = 'disabled')
+
 
     def delete_selected(self):
-        self.db.delete(self.__row_id)
-        self.refresh_list()
+        if messagebox.askokcancel("Verify", "Ok to delete the selected record?"):
+            self.db.delete(self.__row_id)
+            self.refresh_list()
+            self.button_delete_selected.config(state = 'disabled')
      
 
     def clear(self):
